refactor: device’s class and util’s files created
